do not impose send + sync for schedule function for spawn_local
There is a version of spawn, called spawn_local, that does not require
the future to implement Send + Sync. That makes sense, since those
futures are supposed to be used in the same thread in which they were
created.

However, the Send + Sync requirement is still imposed in the schedule
function that accompanies the future implementation. This causes the
future implementation itself to resort to expensive atomic operations
where simple reference counts would do.

To achieve this, I propose the Send+Sync requirement to be removed
from the RawTask, which is an internal abstraction invisible to the
user of this crate.

The Send + Sync requirement on the schedule function will still be
upheld for the function passed to spawn.